+++ packages/editor-common/web/src/Utils/overflowUtils.ts
@@ -27,16 +27,23 @@ export const elementOverflowWithEditor = (element: HTMLElement): Record<string, 
   const elementOffsetTop = elementCoordsInWindow.top;
   const elementWidth = element.clientWidth;
   const elementHeight = element.clientHeight;
-  const rootEditor = element.closest('[data-hook=ricos-editor]') as HTMLElement;
-  const rootEditorCoordsInWindow = rootEditor && getElementCoordsInWindow(rootEditor);
-  const rootEditorOffsetLeft = rootEditorCoordsInWindow.left;
-  const rootEditorOffsetTop = rootEditorCoordsInWindow.top;
-  const editorWidth = (rootEditor && rootEditor.getBoundingClientRect().width) || 999999;
-  const editorHeight = (rootEditor && rootEditor.getBoundingClientRect().height) || 999999;
-  const overflowRight = elementWidth + (elementOffsetLeft - rootEditorOffsetLeft) - editorWidth;
-  const overflowBottom = elementHeight + (elementOffsetTop - rootEditorOffsetTop) - editorHeight;
-  return {
-    overflowRight: overflowRight > 0 ? overflowRight : 0,
-    overflowBottom: overflowBottom > 0 ? overflowBottom : 0,
-  };
+  const rootEditor = element.closest('[data-hook=root-editor]') as HTMLElement;
+  if (rootEditor) {
+    const rootEditorCoordsInWindow = rootEditor && getElementCoordsInWindow(rootEditor);
+    const rootEditorOffsetLeft = rootEditorCoordsInWindow.left;
+    const rootEditorOffsetTop = rootEditorCoordsInWindow.top;
+    const editorWidth = (rootEditor && rootEditor.getBoundingClientRect().width) || 999999;
+    const editorHeight = (rootEditor && rootEditor.getBoundingClientRect().height) || 999999;
+    const overflowRight = elementWidth + (elementOffsetLeft - rootEditorOffsetLeft) - editorWidth;
+    const overflowBottom = elementHeight + (elementOffsetTop - rootEditorOffsetTop) - editorHeight;
+    return {
+      overflowRight: overflowRight > 0 ? overflowRight : 0,
+      overflowBottom: overflowBottom > 0 ? overflowBottom : 0,
+    };
+  } else {
+    return {
+      overflowRight: 0,
+      overflowBottom: 0,
+    };
+  }
 };