+++ packages/editor-common/web/src/consts.ts
@@ -113,6 +113,7 @@ export const INSERT_PLUGIN_BUTTONS = Object.freeze({
   UNDO: 'UndoPlugin_InsertButton',
   REDO: 'RedoPlugin_InsertButton',
   TABLE: 'TablePlugin_InsertButton',
+  COLLAPSIBLE_LIST: 'CollapsibleList_InsertButton',
 });
 
 export const BUTTON_TYPES = Object.freeze({