+++ packages/editor/web/src/RichContentEditor/RichContentEditor.tsx
@@ -30,10 +30,7 @@ import {
   SelectionState,
   setSelectionToBlock,
   emptyDraftContent,
-  EditorModals,
-  getModalStyles,
 } from 'wix-rich-content-editor-common';
-import { getPluginMenuTheme } from './Toolbars/SideToolbar/utils';
 import { convertFromRaw, convertToRaw } from '../../lib/editorStateConversion';
 import { EditorProps as DraftEditorProps, DraftHandleValue } from 'draft-js';
 import { createUploadStartBIData, createUploadEndBIData } from './utils/mediaUploadBI';
@@ -231,8 +228,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
   EditorCommands!: EditorCommands;
 
   getSelectedText!: (editorState: EditorState) => string;
-
-  pluginButtons;
 
   static defaultProps: Partial<RichContentEditorProps> = {
     config: {},
@@ -513,7 +508,6 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
       commonPubsub: this.commonPubsub,
     });
 
-    this.pluginButtons = pluginButtons;
     this.initEditorToolbars(pluginButtons, pluginTextButtons, externalizedButtonProps);
     this.pluginKeyBindings = initPluginKeyBindings(pluginTextButtons);
     this.plugins = [...pluginInstances, ...Object.values(this.toolbars)];
@@ -557,37 +551,9 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
       this.props.textToolbarType === 'static' ? this.toolbars[TOOLBARS.STATIC].Toolbar : null,
   });
 
-  openMobileAddPlugin = () => {
-    const { theme, t, helpers = {}, isMobile, config } = this.props;
-    const addPluginMenuConfig = config
-      ?.getToolbarSettings?.({ textButtons: undefined })
-      .filter(toolbar => toolbar.name === TOOLBARS.MOBILE)[0].addPluginMenuConfig;
-    const structure = this.pluginButtons?.filter(({ buttonSettings }) =>
-      buttonSettings.toolbars.includes(TOOLBARS.MOBILE)
-    );
-    helpers?.openModal?.({
-      modalName: EditorModals.MOBILE_ADD_PLUGIN,
-      modalStyles: getModalStyles({ fullScreen: false, isMobile: true, stickyButtomMobile: true }),
-      plugins: structure.map(({ component, buttonSettings: { name, section } }) => ({
-        component,
-        name,
-        section: section || 'BlockToolbar_Section_Basic',
-      })),
-      theme: getPluginMenuTheme(theme, isMobile),
-      hidePopup: helpers?.closeModal,
-      getEditorState: this.getEditorState,
-      setEditorState: this.setEditorState,
-      pubsub: this.commonPubsub,
-      t,
-      isMobile,
-      addPluginMenuConfig,
-      toolbarName: TOOLBARS.SIDE,
-    });
-  };
-
-  getInnerRCERenderedIn = () => {
+  isInnerRCERenderedInTable = () => {
     const { innerRCERenderedIn } = this.props;
-    return innerRCERenderedIn;
+    return innerRCERenderedIn === 'wix-rich-content-plugin-table';
   };
 
   getInitialEditorState() {
@@ -924,6 +890,10 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     const mode = shouldEnable ? 'render' : 'edit';
     this.editor?.setMode(mode);
     this.inPluginEditingMode = shouldEnable;
+    this.handleSideToolbar(shouldEnable);
+  };
+
+  handleSideToolbar = (shouldEnable: boolean) => {
     const { toolbarsToIgnore: currentToolbarsToIgnore } = this.state;
     const toolbarsToIgnore: ToolbarsToIgnore = currentToolbarsToIgnore;
     const index = toolbarsToIgnore.indexOf('SideToolbar');