+++ packages/plugin-link-preview/web/src/editor.ts
@@ -2,12 +2,15 @@ import { createLinkPreviewPlugin } from './createLinkPreviewPlugin';
 import { LINK_PREVIEW_TYPE, LinkPreviewPluginEditorConfig } from './types';
 import { DEFAULTS } from './defaults';
 import { EditorPluginCreator } from 'wix-rich-content-common';
+import { createRicosExtensions } from './tiptap';
 
 export const pluginLinkPreview: EditorPluginCreator<LinkPreviewPluginEditorConfig> = config => {
+  const pluginConfig = { ...DEFAULTS.config, ...config };
   return {
-    config: { ...DEFAULTS.config, ...config },
+    config: pluginConfig,
     type: LINK_PREVIEW_TYPE,
     createPlugin: createLinkPreviewPlugin,
     ModalsMap: {},
+    tiptapExtensions: createRicosExtensions(pluginConfig),
   };
 };