+++ packages/ricos-common/web/src/RicosEngine.tsx
@@ -97,6 +97,7 @@ export class RicosEngine extends Component<EngineProps> {
       textAlignment,
       onAtomicBlockFocus,
       experiments,
+      iframeSandboxDomain,
     } = this.props;
 
     const { strategyProps, previewContent, htmls } = this.runStrategies();
@@ -176,6 +177,7 @@ export class RicosEngine extends Component<EngineProps> {
       textAlignment,
       onAtomicBlockFocus,
       experiments,
+      iframeSandboxDomain,
     };
 
     const mergedRCProps = merge(strategyProps, _rcProps, ricosPropsToMerge, children.props);