+++ packages/ricos-content/web/src/converters/draft/toDraft/toDraft.ts
@@ -29,6 +29,7 @@ const convert = (ricosContent: RichContent): DraftContent => {
     entityMap: {},
   };
   let latestEntityKey = -1;
+  const usedKeys: Record<string, boolean> = {};
 
   const parseNodes = (index = 0) => {
     const node = nodes[index];
@@ -128,6 +129,10 @@ const convert = (ricosContent: RichContent): DraftContent => {
       },
       blockProps
     );
+    if (usedKeys[newBlock.key] && newBlock.key !== '') {
+      throw Error(`ERROR! Duplicate block key "${newBlock.key}"!`);
+    }
+    usedKeys[newBlock.key] = true;
     draftContent.blocks = [...draftContent.blocks, newBlock];
   };