+++ packages/ricos-editor/web/src/toolbars/TextFormattingToolbar.tsx
@@ -32,18 +32,15 @@ interface TextFormattingToolbarProps {
   linkPanelSettings?: LinkPanelSettings;
   linkSettings?: LinkSettings;
   onInlineToolbarOpen?: (toolbarType: ToolbarType) => void;
-  onToolbarButtonClick?: (name: string, toolbarType: ToolbarType, value?: any) => void;
-  openMobileAddPlugin?: () => void;
+  onToolbarButtonClick?: (
+    name: string,
+    toolbarType: ToolbarType,
+    value?: any,
+    pluginId?: string
+  ) => void;
 }
 
-interface State {}
-
-class TextFormattingToolbar extends Component<TextFormattingToolbarProps, State> {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
+class TextFormattingToolbar extends Component<TextFormattingToolbarProps> {
   getPluginsKey = () => {
     const { activeEditor } = this.props;
     const rawPlugins = activeEditor?.getPlugins?.();
@@ -60,7 +57,6 @@ class TextFormattingToolbar extends Component<TextFormattingToolbarProps, State>
       theme,
       locale,
       getToolbarSettings = () => [],
-      openMobileAddPlugin,
     } = this.props;
     const editorCommands: EditorCommands = activeEditor.getEditorCommands();
     const selection = editorCommands.getSelection();
@@ -107,8 +103,8 @@ class TextFormattingToolbar extends Component<TextFormattingToolbarProps, State>
     const baseStyles = { flex: 'none' };
     const baseMobileStyles = { ...baseStyles, position: 'sticky', top: 0, zIndex: 9 };
     const onInlineToolbarOpen = () => this.props.onInlineToolbarOpen?.(ToolbarType.FORMATTING);
-    const onToolbarButtonClick = (name, value = undefined) => {
-      this.props.onToolbarButtonClick?.(name, ToolbarType.FORMATTING, value);
+    const onToolbarButtonClick = (name, value = undefined, pluginId = undefined) => {
+      this.props.onToolbarButtonClick?.(name, ToolbarType.FORMATTING, value, pluginId);
     };
     const ToolbarToRender = (
       <RicosToolbar
@@ -121,7 +117,6 @@ class TextFormattingToolbar extends Component<TextFormattingToolbarProps, State>
         linkPanelData={linkPanelData}
         colorPickerData={colorPickerData}
         onToolbarButtonClick={onToolbarButtonClick}
-        openMobileAddPlugin={openMobileAddPlugin}
       />
     );
     const ToolbarContainer =