+++ packages/toolbars-new/web/src/modals/link/LinkComponents/BasicLinkPanel.jsx
@@ -2,31 +2,16 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import LinkPanelWrapper from './LinkPanelWrapper';
-// import FocusManager from '../FocusManager';
+import { FocusManager } from 'wix-rich-content-ui-components';
 import { mergeStyles } from 'wix-rich-content-common';
 import styles from '../link-panel.scss';
 import LinkActionsButtons from './LinkActionsButtons';
-import { LinkIcon } from '../Icons';
 
 class BasicLinkPanel extends PureComponent {
   constructor(props) {
     super(props);
     this.styles = mergeStyles({ styles, theme: props.theme });
   }
-
-  renderMobileTitle = () => {
-    const { t } = this.props;
-    return (
-      <div className={styles.mobileLinkModal_titleContainer}>
-        <div className={styles.mobileLinkModal_linkIconContainer}>
-          <LinkIcon />
-        </div>
-        <h3 id="mob_link_modal_hdr" className={styles.mobileLinkModal_title}>
-          {t('MobileLinkModal_Title')}
-        </h3>
-      </div>
-    );
-  };
 
   render() {
     const { styles } = this;
@@ -40,33 +25,35 @@ class BasicLinkPanel extends PureComponent {
       linkPanelValues,
       onChangeLinkPanel,
       isMobile,
+      hasCheckboxes = true,
     } = this.props;
     const linkPanelContainerClassName = classNames(styles.linkPanel_container, {
       [styles.linkPanel_container_isMobile]: isMobile,
+      [styles.basicPanel]: !hasCheckboxes,
     });
+    const showSeparator = hasCheckboxes || isMobile;
     return (
-      // <FocusManager
-      //   className={linkPanelContainerClassName}
-      //   data-hook="linkPanelContainer"
-      //   role="form"
-      //   {...ariaProps}
-      // >
-      <div className={linkPanelContainerClassName} data-hook="linkPanelContainer">
-        {isMobile && this.renderMobileTitle()}
-        <div className={styles.linkPanel_content}>
+      <FocusManager
+        className={linkPanelContainerClassName}
+        data-hook="linkPanelContainer"
+        role="form"
+        {...ariaProps}
+        style={{ padding: 20 }}
+      >
+        <div className={styles.linkPanel_wrapper}>
           <LinkPanelWrapper
             linkValues={linkPanelValues}
             onChange={onChangeLinkPanel}
             showNewTabCheckbox={showNewTabCheckbox}
             showNoFollowCheckbox={showNoFollowCheckbox}
             showSponsoredCheckbox={showSponsoredCheckbox}
+            hasCheckboxes={hasCheckboxes}
             {...sharedPanelsProps}
           />
-          <div className={styles.linkPanel_actionsDivider} role="separator" />
         </div>
-        <LinkActionsButtons basicLinkPanel {...buttonsProps} />
-      </div>
-      // </FocusManager>
+        {showSeparator && <div className={styles.linkPanel_actionsDivider} role="separator" />}
+        <LinkActionsButtons basicLinkPanel {...buttonsProps} basicDisplay={!hasCheckboxes} />
+      </FocusManager>
     );
   }
 }
@@ -84,6 +71,7 @@ BasicLinkPanel.propTypes = {
   linkPanelValues: PropTypes.object,
   onChangeLinkPanel: PropTypes.func,
   isMobile: PropTypes.bool,
+  hasCheckboxes: PropTypes.bool,
 };
 
 export default BasicLinkPanel;