+++ packages/toolbars-new/web/src/modals/link/LinkComponents/MultiSelectLinkPanel.jsx
@@ -5,12 +5,10 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import LinkPanelWrapper from './LinkPanelWrapper';
 import AnchorPanel from '../AnchorComponents/AnchorPanel';
-// import FocusManager from '../FocusManager';
 import { mergeStyles } from 'wix-rich-content-common';
-import { RadioGroup } from 'wix-rich-content-ui-components';
+import { RadioGroup, FocusManager } from 'wix-rich-content-ui-components';
 import styles from '../multi-select-link-panel.scss';
 import LinkActionsButtons from './LinkActionsButtons';
-import { LinkIcon } from '../Icons';
 import { RADIO_GROUP_VALUES } from '../AnchorComponents/consts';
 
 class MultiSelectLinkPanel extends PureComponent {
@@ -43,6 +41,18 @@ class MultiSelectLinkPanel extends PureComponent {
     );
   };
 
+  renderSeparator = () => {
+    const { isMobile } = this.props;
+    return (
+      <div
+        className={classNames(styles.multiSelectLinkPanel_actionsDivider, {
+          [styles.multiSelectLinkPanel_actionsDivider_mobile]: isMobile,
+        })}
+        role="separator"
+      />
+    );
+  };
+
   renderDesktopHeader = () => {
     const { t } = this.props;
     return (
@@ -50,33 +60,23 @@ class MultiSelectLinkPanel extends PureComponent {
         <div className={styles.multiSelectLinkPanel_header}>
           <div>{t('LinkTo_Modal_Header')}</div>
         </div>
-        <div className={styles.multiSelectLinkPanel_actionsDivider} role="separator" />
+        {this.renderSeparator()}
       </>
     );
   };
 
-  renderDesktopLinkActionsButtons = () => {
-    const { buttonsProps } = this.props;
+  renderMultiSelectActions = () => {
+    const { isMobile, buttonsProps } = this.props;
     return (
-      <>
-        <div className={styles.multiSelectLinkPanel_actionsDivider} role="separator" />
+      <div className={isMobile && styles.multiSelectLinkPanel_actionsWrapper_mobile}>
         <LinkActionsButtons {...buttonsProps} />
-      </>
-    );
-  };
-
-  renderMobileHeader = () => {
-    const { buttonsProps, t } = this.props;
-    return (
-      <>
-        <LinkActionsButtons {...buttonsProps} />
-        <div className={styles.multiSelectLinkPanel_header_mobile}>
-          <LinkIcon className={styles.multiSelectLinkPanel_mobileHeaderIcon} />
-          <div>{t('LinkTo_Modal_Header')}</div>
-        </div>
-
-        {this.renderMobileTabs()}
-      </>
+        {isMobile && (
+          <div>
+            {this.renderSeparator()}
+            {this.renderMobileTabs()}
+          </div>
+        )}
+      </div>
     );
   };
 
@@ -124,6 +124,7 @@ class MultiSelectLinkPanel extends PureComponent {
       anchorableBlocksData,
       anchorPanelValues,
       blockPreview,
+      hasCheckboxes = true,
     } = this.props;
 
     const containerClassName = !isMobile
@@ -139,14 +140,14 @@ class MultiSelectLinkPanel extends PureComponent {
       : styles.multiSelectLinkPanel_LinkPanelContainer_mobile;
 
     return (
-      // <FocusManager
-      //   className={containerClassName}
-      //   data-hook="linkPanelContainer"
-      //   role="form"
-      //   {...ariaProps}
-      // >
-      <div className={containerClassName} data-hook="linkPanelContainer">
-        {!isMobile ? this.renderDesktopHeader() : this.renderMobileHeader()}
+      <FocusManager
+        className={containerClassName}
+        data-hook="linkPanelContainer"
+        role="form"
+        {...ariaProps}
+        style={{ padding: 20 }}
+      >
+        {!isMobile && this.renderDesktopHeader()}
 
         <div className={contentClassName}>
           {!isMobile && (
@@ -163,6 +164,7 @@ class MultiSelectLinkPanel extends PureComponent {
                 showNewTabCheckbox={showNewTabCheckbox}
                 showNoFollowCheckbox={showNoFollowCheckbox}
                 showSponsoredCheckbox={showSponsoredCheckbox}
+                hasCheckboxes={hasCheckboxes}
                 {...sharedPanelsProps}
               />
             </div>
@@ -177,10 +179,9 @@ class MultiSelectLinkPanel extends PureComponent {
             />
           )}
         </div>
-
-        {!isMobile && this.renderDesktopLinkActionsButtons()}
-      </div>
-      //</FocusManager>
+        {!isMobile && this.renderSeparator()}
+        {this.renderMultiSelectActions()}
+      </FocusManager>
     );
   }
 }
@@ -202,6 +203,7 @@ MultiSelectLinkPanel.propTypes = {
   anchorableBlocksData: PropTypes.object,
   anchorPanelValues: PropTypes.object,
   isMobile: PropTypes.bool,
+  hasCheckboxes: PropTypes.bool,
   blockPreview: PropTypes.func,
 };