+++ CHANGELOG.md
@@ -17,6 +17,33 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+
+- `plugin-common`
+  - [#2842](https://github.com/wix/ricos/pull/2842) fix link toolbar shortcut
+
+</details>
+<hr/>
+
+## 8.48.4 (Aug 10, 2021)
+
+### :bug: Bug Fix
+
+- `collapsible-list`
+
+  - [#2831](https://github.com/wix/ricos/pull/2831) add collapsible_list to INSERT_PLUGIN_BUTTONS
+
+- `ricos-viewer` `ricos-editor`
+  - [#2840](https://github.com/wix/ricos/pull/2840) expose `iframeSandboxDomain` prop on editor & viewer fix
+  -
+
+## 8.48.3 (Aug 09, 2021)
+
+### :bug: Bug Fix
+
+- `link`
+  - [#2830](https://github.com/wix/ricos/pull/2830) fix modal design when opened using image toolbar
+
 ### :nail_care: Polish
 
 - `ricos-viewer` `ricos-editor`
@@ -26,14 +53,6 @@
 
 - `editor-common`
   - [#2824](https://github.com/wix/ricos/pull/2824) overflowUtils - support overflow from bottom + change returned value
-
-### :bug: Bug Fix
-
-- `link`
-  - [#2830](https://github.com/wix/ricos/pull/2830) fix modal design when opened using image toolbar
-
-</details>
-<hr/>
 
 ## 8.48.2 (Aug 05, 2021)
 
@@ -61,6 +80,8 @@
 
 ### :bug: Bug Fix
 
+- `ricos-common`
+  - [#2831](https://github.com/wix/ricos/pull/2831) add collapsible_list plugin to INSERT_PLUGIN_BUTTONS
 - `editor`
   - [#2813](https://github.com/wix/ricos/pull/2813) fix rel value not applying in link panel by config